[23] Update Elements#getDocComment() to support markdown #3189
